refactor: update loan-related DTOs and services to use Double for interest rates and streamline loan creation process

+++ server/src/main/java/com/lankacapital/server/dtos/InterestUpdateDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class InterestUpdateDTO {
     private String fileNumber;
-    private Integer interestRate;
+    private Double interestRate;
 }
